blog/requirements.txt
dj-database-url==0.5.0
Django==2.2.2
django-heroku==0.3.1
djangorestframework==3.9.4